+++ src/components/Launch.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import {View, Image, StyleSheet} from 'react-native';
 import {width, height, k} from './Global';
+import * as Progress from 'react-native-progress';
+import {colors} from '../constants';
 
 export default class Launch extends React.Component {
   render() {
     return (
       <View style={{flex: 1, alignItems: 'center', backgroundColor: 'transparent'}}>
         <View style={styles.container}>
-          <Image style={styles.backgroundImage} source={require('../../images/EmptyBackground.png')} />
+          <Image style={styles.backgroundImage} source={require('../../images/Launch.gif')} />
         </View>
-        <View style={styles.container}>
-          <Image style={{width: 117, height: 117}} source={require('../../images/loading.gif')} />
-        </View>
+        <Progress.CircleSnail size={26 * k} thickness={2} style={{top: 556 * k}} color={colors.PINK} />
         {this.props.children}
       </View>
     );