+++ src/components/Logo.js
@@ -7,7 +7,7 @@ export default class extends React.Component {
   render() {
     return (
       <View style={styles.container}>
-        <Image style={{width: 320, height: 568}} source={require('../../images/LaunchScreen.png')} />
+        <Image style={{width: 320, height: 568}} source={require('../../images/Launch.gif')} />
       </View>
     );
   }