+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Ticket numbers refer to the ticket tracker for this project if not specified.
 * Fix: User is unable to tap on their own avatar to see their own profile (#711)
 * Fix: Within Bot Profile, username is not tappable (#799)
 * Fix: User gets stuck on Explore Nearby while playing on Explore Nearby (#825)
+* Onboarding: Loading Screen (#802)
 * Onboarding: Configure Twitter Digits Design Flow (#813)
 * User Profile: Metadata Bar & other user details (#667)